Use GODOG's --format junit option to produce jUnit output from the tests.
However, in v0.9.0 output to file is not implemented yet -- coming in v0.11.0 and already available in v0.11.0-rc1. Cannot use that version of godog because it does not support BeforeFeature hooks. This branch modifies the Makefile in such a way that `make integration` runs the dummy.feature and prints the output in the jUnit format to stdout.