gha: unify all windows artifacts builds into one workflow
This helps circumvent multiple complications. One, windows-e2e
workflow does not need to wait for 2 workflows to finish (which is
hard for a workflow_run triggered workflow). Two, we do not need to
worry about matching the right artifacts from builder workflows with
the right workflow_run that needs them (formerly used commit sha to do
this). Third, the organization of workflows becomes logical: builder
that prepares artifacts, and tester that consumes them and runs tests.